Removed Inverse Electrolyzer 1 Input
