+++ Source/InverseElectrolyzer/InverseElectrolyzer.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{66B47D3F-2308-4E0B-A1E7-57386711F1F5}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>InverseElectrolyzerMod</RootNamespace>
-    <AssemblyName>InverseElectrolyzerMod</AssemblyName>
+    <RootNamespace>InverseElectrolyzer</RootNamespace>
+    <AssemblyName>InverseElectrolyzer</AssemblyName>
     <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -25,7 +25,7 @@
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>none</DebugType>
     <Optimize>true</Optimize>
-    <OutputPath>..\..\Mods\InverseElectrolyzerMod\</OutputPath>
+    <OutputPath>..\..\Mods\InverseElectrolyzer\</OutputPath>
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -64,21 +64,15 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <None Include="InverseElectrolyzerConfig.cs.bak" />
-    <None Include="InverseElectrolyzer.cs.bak" />
-    <Compile Include="InverseElectrolyzerAlt.cs" />
-    <Compile Include="InverseElectrolyzerAltConfig.cs" />
-    <Compile Include="InverseElectrolyzerMod.cs" />
     <Compile Include="InverseElectrolyzer.cs" />
     <Compile Include="InverseElectrolyzerConfig.cs" />
+    <Compile Include="InverseElectrolyzerMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>
-xcopy "$(TargetDir.TrimEnd('\'))" "Z:\Applications\Steam Games\SteamApps\common\OxygenNotIncluded\Mods\$(ProjectName)" /Y /S /E /I
-
-xcopy "$(TargetDir.TrimEnd('\'))" "C:\Program Files (x86)\Steam\SteamApps\common\OxygenNotIncluded\Mods\$(ProjectName)" /Y /S /E /I
+xcopy "$(TargetDir.TrimEnd('\'))" "%25UserProfile%25\Documents\Klei\OxygenNotIncluded\mods\dev\$(ProjectName)" /Y /S /E /I
 </PostBuildEvent>
   </PropertyGroup>
 </Project>
