+++ Source/InverseElectrolyzer/Properties/AssemblyInfo.cs
