+++ Source/ModLoader.sln
@@ -20,7 +20,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "10_Experimental", "10_Exper
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InstantResearch", "InstantResearch\InstantResearch.csproj", "{8268151B-2A54-4539-8CC2-85F0494AA04A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InverseElectrolyzerMod", "InverseElectrolyzerMod\InverseElectrolyzerMod.csproj", "{66B47D3F-2308-4E0B-A1E7-57386711F1F5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InverseElectrolyzer", "InverseElectrolyzer\InverseElectrolyzer.csproj", "{66B47D3F-2308-4E0B-A1E7-57386711F1F5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "11_Outdated", "11_Outdated", "{05AEF266-9A07-41B4-83C5-FC6632B27A0E}"
 EndProject
@@ -313,7 +313,7 @@ Global
 		{3DDC5DF0-0321-41EA-B556-022CEE509815} = {0263B9CC-A5D0-433C-82D2-20AAD99BAE40}
 		{F23F1C19-7F80-41CC-86F7-F720C6D3C43B} = {0263B9CC-A5D0-433C-82D2-20AAD99BAE40}
 		{8268151B-2A54-4539-8CC2-85F0494AA04A} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
-		{66B47D3F-2308-4E0B-A1E7-57386711F1F5} = {3DDC5DF0-0321-41EA-B556-022CEE509815}
+		{66B47D3F-2308-4E0B-A1E7-57386711F1F5} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
 		{05AEF266-9A07-41B4-83C5-FC6632B27A0E} = {0263B9CC-A5D0-433C-82D2-20AAD99BAE40}
 		{DB33F126-AECD-4E02-9F1F-1E4B08DC27C4} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
 		{74B7B353-BB3E-4EE1-9035-7357D5472D39} = {3DDC5DF0-0321-41EA-B556-022CEE509815}
